requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
python-multipart
sqlalchemy
pydantic
aiofiles
# PDF Parsing
PyMuPDF

# DOCX Parsing
python-docx